+++ src/modules/index.ts
@@ -1,3 +1,3 @@
-import PrayerTimesModule from "./prayer-times/prayer-times";
+import PrayerTimesModule from "./prayer-times/prayer-times.module";
 
 export default {PrayerTimesModule};
